+++ inc/texts.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * This file only exists to make individual text abbreviations available to the translation tool.
- *
- * @package personio-integration-light
- */
-
-esc_attr_e( 'success', 'personio-integration-light' );
-esc_attr_e( 'error', 'personio-integration-light' );